added asteroid class to objects.
